refactor compose cli commands into separate file
